requirements.txt
pymupdf>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.24
pydantic>=2.6
typer>=0.12
rich>=13.7
orjson>=3.10
tqdm>=4.66
fastapi>=0.110
uvicorn>=0.29
transformers>=4.41
torch>=2.1
